numpy.convert: conversion between RGB formats, and between YUV formats, but not between the two. No v210
